Extract the SQLite3GlobalState to a separate class to fix a JavaDoc issue
