+++ src/sqlancer/sqlite3/gen/SQLite3CreateVirtualRtreeTabelGenerator.java
@@ -7,7 +7,7 @@ import sqlancer.Randomly;
 import sqlancer.common.DBMSCommon;
 import sqlancer.common.query.ExpectedErrors;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 import sqlancer.sqlite3.schema.SQLite3Schema.SQLite3Column;
 
 public final class SQLite3CreateVirtualRtreeTabelGenerator {
