+++ src/sqlancer/sqlite3/gen/SQLite3ExplainGenerator.java
@@ -2,9 +2,9 @@ package sqlancer.sqlite3.gen;
 
 import sqlancer.Randomly;
 import sqlancer.common.query.SQLQueryAdapter;
+import sqlancer.sqlite3.SQLite3GlobalState;
 import sqlancer.sqlite3.SQLite3Provider;
 import sqlancer.sqlite3.SQLite3Provider.Action;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
 
 public final class SQLite3ExplainGenerator {
