+++ src/sqlancer/sqlite3/gen/SQLite3PragmaGenerator.java
@@ -6,7 +6,7 @@ import java.util.function.Supplier;
 import sqlancer.Randomly;
 import sqlancer.common.query.ExpectedErrors;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 
 public class SQLite3PragmaGenerator {
