+++ src/sqlancer/sqlite3/gen/SQLite3VacuumGenerator.java
@@ -3,7 +3,7 @@ package sqlancer.sqlite3.gen;
 import sqlancer.Randomly;
 import sqlancer.common.query.ExpectedErrors;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 
 /**
  * @see <a href="https://www.sqlite.org/lang_vacuum.html">VACUUM</a>
