+++ src/sqlancer/sqlite3/gen/ddl/SQLite3DropIndexGenerator.java
@@ -3,7 +3,7 @@ package sqlancer.sqlite3.gen.ddl;
 import sqlancer.Randomly;
 import sqlancer.common.query.ExpectedErrors;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 
 // see https://www.sqlite.org/lang_dropindex.html
 public final class SQLite3DropIndexGenerator {
