+++ src/sqlancer/sqlite3/gen/ddl/SQLite3DropTableGenerator.java
@@ -4,7 +4,7 @@ import sqlancer.IgnoreMeException;
 import sqlancer.Randomly;
 import sqlancer.common.query.ExpectedErrors;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 
 public final class SQLite3DropTableGenerator {
