+++ src/sqlancer/sqlite3/oracle/tlp/SQLite3TLPWhereOracle.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import sqlancer.ComparatorHelper;
 import sqlancer.Randomly;
-import sqlancer.sqlite3.SQLite3Provider.SQLite3GlobalState;
+import sqlancer.sqlite3.SQLite3GlobalState;
 import sqlancer.sqlite3.SQLite3Visitor;
 
 public class SQLite3TLPWhereOracle extends SQLite3TLPBase {
